Tests/LibWeb/Text/expected/HTML/Window-find-no-match-preserves-selection.txt

window.find("findable"): true, rangeCount: 1
window.find("nonexistent"): false, rangeCount: 1
